pyproject.toml
[tool.poetry]
name = "Task4"
version = "0.1.0"
description = ""
authors = ["Troy,Li Ting, Aaron"]

[tool.poetry.dependencies]
python = "^3.12.0"
dvc = "^2.10.0"
hydra-core = "1.3.2"
pdoc3 = "^0.10.0"
cython = "3.0.8"
flask = "3.0.2"
gunicorn = "21.2.0"
cookiecutter = "2.6.0"
werkzeug = "3.0.1"
wtforms = "3.1.2"
flask-wtf = "1.2.1"
ipython = "8.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"
ipykernel = "^6.28.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E501"]
select = ["B","C","E","F","W","B9", "I", "Q"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.mccabe]
max-complexity = 10
